requirements_workshop.txt
# Pixels2GenAI Workshop Requirements
# ===================================
# Minimal dependencies for the 6-module workshop:
# - 1.1.1 RGB Basics
# - 1.2.2 Cellular Automata (Game of Life)
# - 3.4.1 Convolution
# - 4.1.1 Fractal Square
# - 9.1.1 Perceptron from Scratch
# - 12.3.1 DDPM Basics
#
# Install with: pip install -r requirements_workshop.txt

# Core (all modules)
numpy>=1.21.0
pillow>=8.3.0

# Cellular Automata & Convolution (scipy.ndimage.convolve)
scipy>=1.7.0

# Visualization & Animation
matplotlib>=3.4.0
imageio>=2.9.0

# DDPM Module (12.3.1)
torch>=1.9.0
tqdm>=4.62.0
denoising-diffusion-pytorch>=1.0.0
